Fix test for WordPress version 4.3
